+++ lib/yggdrasil/settings.ex
@@ -4,23 +4,6 @@ defmodule Yggdrasil.Settings do
   """
   use Skogsra
 
-  #########
-  # Helpers
-
-  @doc false
-  def gen_env_name(namespace, key, default \\ "_YGGDRASIL_") do
-    prefix =
-      namespace
-      |> Atom.to_string()
-      |> String.trim("Elixir.")
-      |> String.upcase()
-      |> String.split(".")
-      |> Enum.join("_")
-
-    name = key |> Atom.to_string() |> String.upcase()
-    prefix <> default <> name
-  end
-
   ############################
   # Yggdrasil process registry
 
@@ -32,13 +15,9 @@ defmodule Yggdrasil.Settings do
     process_registry: ExReg
   ```
   """
-  app_env(
-    :yggdrasil_process_registry,
-    :yggdrasil,
-    :process_registry,
+  app_env :yggdrasil_process_registry, :yggdrasil, :process_registry,
     skip_system: true,
     default: ExReg
-  )
 
   ############################
   # Yggdrasil registry options
@@ -51,13 +30,9 @@ defmodule Yggdrasil.Settings do
     module_registry: :yggdrasil_registry
   ```
   """
-  app_env(
-    :yggdrasil_module_registry,
-    :yggdrasil,
-    :module_registry,
+  app_env :yggdrasil_module_registry, :yggdrasil, :module_registry,
     skip_system: true,
     default: :yggdrasil_registry
-  )
 
   ################################
   # Yggdrasil distribution options
@@ -70,13 +45,9 @@ defmodule Yggdrasil.Settings do
     pubsub_adapter: Phoenix.PubSub.PG2
   ```
   """
-  app_env(
-    :yggdrasil_pubsub_adapter,
-    :yggdrasil,
-    :pubsub_adapter,
+  app_env :yggdrasil_pubsub_adapter, :yggdrasil, :pubsub_adapter,
     skip_system: true,
     default: Phoenix.PubSub.PG2
-  )
 
   @doc """
   Pub-sub name. By default is `Yggdrasil.PubSub`.
@@ -86,13 +57,9 @@ defmodule Yggdrasil.Settings do
     pubsub_name: Yggdrasil.PubSub
   ```
   """
-  app_env(
-    :yggdrasil_pubsub_name,
-    :yggdrasil,
-    :pubsub_name,
+  app_env :yggdrasil_pubsub_name, :yggdrasil, :pubsub_name,
     skip_system: true,
     default: Yggdrasil.PubSub
-  )
 
   @doc """
   Pub-sub options. By default are `[pool_size: 1]`.
@@ -102,13 +69,9 @@ defmodule Yggdrasil.Settings do
     pubsub_options: [pool_size: 1]
   ```
   """
-  app_env(
-    :yggdrasil_pubsub_options,
-    :yggdrasil,
-    :pubsub_options,
+  app_env :yggdrasil_pubsub_options, :yggdrasil, :pubsub_options,
     skip_system: true,
     default: [pool_size: 1]
-  )
 
   #############################
   # Yggdrasil publisher options
@@ -122,11 +85,7 @@ defmodule Yggdrasil.Settings do
     publisher_options: [size: 5, max_overflow: 10]
   ```
   """
-  app_env(
-    :yggdrasil_publisher_options,
-    :yggdrasil,
-    :publisher_options,
+  app_env :yggdrasil_publisher_options, :yggdrasil, :publisher_options,
     skip_system: true,
     default: [size: 5, max_overflow: 10]
-  )
 end